BackendDjango/requirements.txt
Django==1.10.5
httpie==0.9.2
ipython==3.0.0
Pygments==2.0.2
requests==2.5.3
django-cors-middleware==1.3.1
django-extensions==1.7.1
djangorestframework==3.4.4
djangorestframework-jwt==1.8.0
drf-nested-routers==0.9.0
PyJWT==1.4.2
six==1.10.0
sendgrid-django
python-dotenv
stripe